-inline void* ToPointer(int32 i) { return _ToPointer(i); } -inline int32 ToInt32(void *p) { return _ToInt32(p); } +inline void* ToPointer(uint32 i) { return _ToPointer(i); } +inline uint32 ToInt32(void *p) { return _ToInt32(p); }
and similar changes


git-svn-id: https://gforge.info.ucl.ac.be/svn/mozart@2951 ada56829-ad1f-0410-b00f-83cda6628aec

+++ trunk/mozart/platform/emulator/codearea.hh
@@ -348,7 +348,7 @@ void displayCode(ProgramCounter from, int ssize);
  */
 
 class InlineCache {
-  int32 key;
+  uint32 key;
   int32 value;
       
 public:
@@ -356,7 +356,7 @@ public:
 
   int lookup(SRecord *rec, TaggedRef feature)
   {
-    if (key!=rec->getSRecordArity()) {
+    if (key!=(uint32)rec->getSRecordArity()) {
       key   = rec->getSRecordArity();
       value = rec->getIndex(feature); // is ok even if index==-1 !
     }
